(Partially) serialize experiment results
Graphs are not serialized, due to the labels on edges being
Java CharPred instances, not lending themselves to the quick
and dirty solution I need right now.

Solution: Float the previous table along, we can create the SFA
from that for inspection, and if need be we can recreate the tree.